Room Messages

+++ base/views.py
@@ -1,16 +1,21 @@
 from django.shortcuts import render, redirect
 from django.contrib import messages
-
+from django.contrib.auth.decorators import login_required
 from django.db.models import Q
 from django.contrib.auth.models import User
 from django.contrib.auth import authenticate, login, logout
 from django.http import HttpResponse
+from django.contrib.auth.forms import UserCreationForm
 from .models import Room, Topic
 from .forms import RoomForm
 
 def loginPage(request):
+    page = 'login'
+    if request.user.is_authenticated:
+        return redirect('home')
+
     if request.method == 'POST':
-        username = request.POST.get('username')
+        username = request.POST.get('username').lower()
         password = request.POST.get('password')
 
         try:
@@ -26,9 +31,29 @@ def loginPage(request):
             messages.error(request, "Username or password does not exists.")
 
 
-    context = {}
+    context = {'page': page}
     return render(request, 'base/login_register.html', context)
 
+def logoutUser(request):
+    logout(request)    
+    return redirect('home')
+    
+def registerPage(request):
+    form = UserCreationForm()
+    if request.method == 'POST':
+        form = UserCreationForm(request.POST)
+
+        if form.is_valid():
+            user = form.save(commit=False)
+            user.username = user.username.lower()
+            user.save()
+            login(request,user)
+            return redirect('home')
+        else:
+            messages.error(request, 'An error occurred during registrtaion')
+
+    context = {'form': form}
+    return render(request, 'base/login_register.html', context) 
 
 def home(request):
     q = request.GET.get('q') if request.GET.get('q') != None else ''
@@ -45,9 +70,11 @@ def home(request):
 
 def room(request, pk):
     room = Room.objects.get(id=pk)
-    context = {'room': room}
+    room_messages = room.message_set.all()
+    context = {'room': room, 'room_messages': room_messages}
     return render(request, 'base/room.html', context)
 
+@login_required(login_url='login')
 def createRoom(request):
     form = RoomForm()
     
@@ -61,9 +88,13 @@ def createRoom(request):
     context = {'form': form}
     return render(request, 'base/room_form.html',context)
 
+@login_required(login_url='login')
 def updateRoom(request, pk):
     room = Room.objects.get(id = pk)
     form = RoomForm(instance=room)
+
+    if request.user != room.host:
+        return HttpResponse('You are not allowed here')
 
     if(request.method == 'POST'):
         form = RoomForm(request.POST, instance=room)
@@ -77,7 +108,9 @@ def updateRoom(request, pk):
 def deleteRoom(request, pk):
 
     room = Room.objects.get(id = pk)
-
+    if request.user != room.host:
+        return HttpResponse('You are not allowed here')
+    
     if(request.method == 'POST'):
         room.delete()
         return redirect('home')
